+++ docs/changes.rst
@@ -8,6 +8,9 @@ To be released.
 
 - Fixed :exc:`TypeError` that rises when :class:`CloudMasterKeyStore
   <geofront.masterkey.CloudMasterKeyStore>` is used with AWS S3 driver.
+- Added :option:`--trusted-proxy <geofront-server --trusted-proxy>` option
+  to :program:`goefront-server` command.  It's useful when the server is
+  run behind a reverse proxy.
 
 
 Version 0.1.0